r23044: - use uint32_t for handle id's - include the session vuid in the SMB2 128-Bit wire handles   as SMB2 oplock breaks doesn't include a TID or VUID in the header   we need to make sure the handle is unique for the whole TCP connection
metze

+++ source/smb_server/smb_server.h
@@ -185,13 +185,14 @@ struct smbsrv_handle {
 	/*
 	 * the value passed over the wire
 	 * - 16 bit for smb
-	 * - 64 bit for smb2
+	 * - 32 bit for smb2
 	 *   Note: for SMB2 handles are 128 bit
-	 *         we'll fill the 2nd 64 bit with:
+	 *         we'll fill them with
+	 *	   - 32 bit HID
 	 *         - 32 bit TID
-	 *         - 32 bit 0xFFFFFFFF
+	 *	   - 64 bit VUID
 	 */
-	uint64_t hid;
+	uint32_t hid;
 
 	/*
 	 * the ntvfs handle passed to the ntvfs backend
